feat: configurable touch gestures
Adds configurable multi-finger touch gesture support with
per-gesture settings for finger count, sensitivity, and
natural scroll. A standalone GTK4/Libadwaita settings app
(niri-touch-settings) is available as a separate project.
